Changed dynamodb.js to skip mapping the committee topic. So it is no longer displayed to the user as an option to subscribe.
topics.model.js was calling formatItems() which was mapping the topics from dynamodb to an array called automated[].
Those topics where then displayed to the user in the view list.pug L42.

This is a temporary fix as currently notifications for committees has been disabled as it was broken and sending lots of redundant emails.